evals/requirements.txt
pandas>=2.0.0
datasets>=2.0.0
numpy>=1.24.0
ragas>=0.2.0
langchain-openai>=0.1.0
langchain-aws>=1.0.0
pyyaml>=6.0.1
boto3>=1.34.0

